Split the refusal diagnosis into its own module
