+++ src/shared/db/attendees/capacity/checks.ts
@@ -13,15 +13,13 @@ import {
 import {
   inPlaceholders,
   queryAll,
-  queryBatchPrimary,
   requireOne,
-  resultRows,
   type SqlStatement,
 } from "#db/client.ts";
 import { listingGroups } from "#db/groups.ts";
 import { getListingWithCount } from "#db/listings/records.ts";
 import { type NumberedSql, numberedStatement } from "#db/numbered-statement.ts";
-import { identity, map, mapById, requiredMapValue, unique } from "#fp";
+import { identity, map, mapById, unique } from "#fp";
 import { capacityDateFor, countsPerDate } from "#shared/capacity-rules.ts";
 import { dateToStartEnd, expandDailyRange } from "./range.ts";
 import type { ListingCapacityRow } from "./types.ts";
@@ -141,7 +139,7 @@ export const checkListingAvailability = async (
   )[0]!;
 };
 
-const getOrCreateBucket = <K>(
+export const getOrCreateBucket = <K>(
   buckets: Map<K, CapacityBucket>,
   key: K,
 ): CapacityBucket => {
@@ -167,7 +165,7 @@ const getOrCreateBucket = <K>(
  * also pins the running total its write statement reads: no undated
  * statement of the write runs after it, so a dated line booked later never
  * raises that state. */
-const addDemandToBucket = (
+export const addDemandToBucket = (
   bucket: CapacityBucket,
   listing: Pick<ListingCapacityRow, "listing_type">,
   item: BatchAvailabilityItem,
@@ -255,189 +253,3 @@ const fitsThrough =
     const row = await read<{ fits: number }>(sql, args);
     return row.fits === 1;
   };
-
-type LineListingFacts = {
-  groupIds: number[];
-  listing_type: ListingCapacityRow["listing_type"];
-};
-
-/** One cart demand for a slice of lines, each line counted on its own date. */
-const linesDemand = (
-  lines: LineBooking[],
-  factsById: Map<number, LineListingFacts>,
-): CartDemand => {
-  const demand: CartDemand = {
-    groupDemand: new Map(),
-    listingDemand: new Map(),
-  };
-  for (const line of lines) {
-    const facts = requiredMapValue(
-      factsById,
-      line.listingId,
-      `Listing ${line.listingId} was not read for the refusal diagnosis`,
-    );
-    const item = {
-      durationDays: line.durationDays,
-      listingId: line.listingId,
-      quantity: line.quantity,
-    };
-    addDemandToBucket(
-      getOrCreateBucket(demand.listingDemand, line.listingId),
-      facts,
-      item,
-      line.date,
-    );
-    for (const groupId of facts.groupIds) {
-      addDemandToBucket(
-        getOrCreateBucket(demand.groupDemand, groupId),
-        facts,
-        item,
-        line.date,
-      );
-    }
-  }
-  return demand;
-};
-
-/** The write's guarded statements run in write order, each seeing the rows the
- * earlier ones inserted. Each prefix of the order is asked as one cumulative
- * demand over that same order, so the first line that does not fit on top of
- * its predecessors is the one named — the statement the write aborted on. A
- * shared group limit counts, whatever dates the lines sit on. Prefix fits
- * only shrink as lines are added, so first unfit prefix is the culprit, and
- * every prefix a batch asks runs inside ONE primary statement batch — the
- * answers of a batch share one snapshot and cannot contradict each other.
- *
- * The reads run on the primary because the refused write did. A replica can lag
- * behind the booking that took the last place, and the isolate's caches can
- * hold a listing another isolate deleted. */
-
-/** Prefixes one probe batch samples. Eight keeps a whole typical order's
- * diagnosis inside ONE snapshot batch; a longer order narrows its bracket by
- * this factor per batch, so the statements per request stay bounded however
- * many lines the order has — one statement per prefix would grow the request
- * with the square of the length. The whole-order probe rides every batch, so
- * a room freed between batches still names no listing. */
-const PROBE_STRIDE = 8;
-
-/** The prefixes one batch asks: evenly by stride from the fitting end of
- * the bracket, plus the bracket ends and the whole order. */
-const batchProbePrefixes = (
-  longestFit: number,
-  shortestUnfit: number,
-  orderLength: number,
-): number[] => {
-  const step = Math.ceil((shortestUnfit - longestFit) / PROBE_STRIDE);
-  const sampled: number[] = [];
-  for (let prefix = longestFit + step; prefix < shortestUnfit; prefix += step) {
-    sampled.push(prefix);
-  }
-  // Ascending, so the bracket walk below finds the first unfit prefix in
-  // numeric order rather than insertion order.
-  return unique([
-    ...sampled,
-    ...(longestFit > 0 ? [longestFit] : []),
-    shortestUnfit,
-    orderLength,
-  ]).sort((left, right) => left - right);
-};
-
-/** Ask one batch of prefixes through one primary snapshot. */
-const askWhetherPrefixesFit = async (
-  lines: LineBooking[],
-  factsById: Map<number, LineListingFacts>,
-  probes: readonly number[],
-): Promise<Map<number, boolean>> => {
-  const results = await queryBatchPrimary(
-    probes.map((prefix) =>
-      buildCartCapacitySql(linesDemand(lines.slice(0, prefix), factsById)),
-    ),
-  );
-  return new Map(
-    probes.map((prefix, index) => [
-      prefix,
-      resultRows<{ fits: number }>(results[index]!)[0]!.fits === 1,
-    ]),
-  );
-};
-
-/** The index of the first line that does not fit, found by stride batches
- * over the order's prefixes: a typical order's whole search is one snapshot
- * batch, a longer one narrows a bracket by the stride factor per batch.
- * Every batch probes the bracket ends it carries: a room freed between
- * batches makes the whole order fit again and names nothing, and a room
- * consumed outruns the fitting bound, whose failure sends the bracket back
- * to the start. A line is named only when one batch proves an adjacent
- * prefix pair — fitting below, unfit above. Null means nothing is proven
- * at a snapshot, and no line is named. */
-const firstUnfitLineIndex = async (
-  lines: LineBooking[],
-  factsById: Map<number, LineListingFacts>,
-): Promise<number | null> => {
-  let longestFit = 0;
-  let shortestUnfit = lines.length;
-  // The budget keeps a room that changes between every batch from spinning
-  // the search forever; running it out names nothing below.
-  for (let batch = 0; batch < lines.length; batch++) {
-    const probes = batchProbePrefixes(longestFit, shortestUnfit, lines.length);
-    const fits = await askWhetherPrefixesFit(lines, factsById, probes);
-    // The whole order fitting again at this snapshot means the room was
-    // freed again and no line is named.
-    if (fits.get(lines.length)) return null;
-    // The smallest probe IS the carried fitting bound when it is above
-    // zero, so a booking that outran it makes it the first unfit prefix and
-    // the bracket falls back to (0, bound] on its own — the next batch
-    // re-metres from the start.
-    const firstUnfit = probes.find((prefix) => !fits.get(prefix))!;
-    const fittingBefore = probes[probes.indexOf(firstUnfit) - 1] ?? 0;
-    if (firstUnfit - fittingBefore <= 1) return firstUnfit - 1;
-    longestFit = fittingBefore;
-    shortestUnfit = firstUnfit;
-  }
-  // The budget ran out with no adjacent probed pair, so the room kept
-  // changing between snapshots and no line is proven the culprit: an
-  // unproven guess would name a line that may fit right now.
-  return null;
-};
-
-export const refusedOrderUnfitListingIds = async (
-  lines: LineBooking[],
-): Promise<number[]> => {
-  if (lines.length === 0) return [];
-  const listingIds = unique(lines.map((line) => line.listingId));
-  const [listingResult, memberResult] = await queryBatchPrimary([
-    {
-      args: listingIds,
-      sql: `SELECT listing.id, listing.listing_type
-              FROM listings AS listing
-             WHERE listing.id IN (${inPlaceholders(listingIds)})`,
-    },
-    {
-      args: listingIds,
-      sql: `SELECT groupListing.listing_id, groupListing.group_id
-              FROM group_listings AS groupListing
-             WHERE groupListing.listing_id IN (${inPlaceholders(listingIds)})`,
-    },
-  ]);
-  const factsById = new Map<number, LineListingFacts>(
-    resultRows<Pick<ListingCapacityRow, "id" | "listing_type">>(
-      listingResult!,
-    ).map((row) => [row.id, { groupIds: [], listing_type: row.listing_type }]),
-  );
-  if (listingIds.some((id) => !factsById.has(id))) return [];
-  for (const row of resultRows<{ group_id: number; listing_id: number }>(
-    memberResult!,
-  )) {
-    requiredMapValue(
-      factsById,
-      row.listing_id,
-      `Group membership row for an unrequested listing ${row.listing_id}`,
-    ).groupIds.push(row.group_id);
-  }
-
-  // The reads run on the primary because the refused write did. A replica
-  // can lag behind the booking that took the last place, and the isolate's
-  // caches can hold a listing another isolate deleted.
-  const firstUnfit = await firstUnfitLineIndex(lines, factsById);
-  return firstUnfit === null ? [] : [lines[firstUnfit]!.listingId];
-};
