+++ test/shared/db/attendees/capacity/refusal-diagnosis/helpers.test.ts
@@ -1,7 +1,7 @@
 import { expect } from "@std/expect";
 import { it as test } from "@std/testing/bdd";
 import { enableQueryLog, runWithQueryLogContext } from "#db/query-log.ts";
-import { awaitObservedProbe } from "./diagnosis-helpers.ts";
+import { awaitObservedProbe } from "./helpers.ts";
 
 test("awaitObservedProbe fails loudly when no probe statement lands", async () => {
   await runWithQueryLogContext(async () => {
