+++ test/shared/db/attendees/capacity/refusal-diagnosis/helpers.ts
